Separate query state and hooks.
